requirements.txt
# Web Framework and Services
fastapi>=0.115.9
uvicorn>=0.34.0
python-multipart>=0.0.20
httpx>=0.28.1
pydantic-settings>=2.8.1

# Database Related
SQLAlchemy>=2.0.38
alembic>=1.14.1
asyncpg>=0.30.0
# psycopg2-binary>=2.9.0

# Authentication and Security
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0

# Asynchronous Task Processing
# celery>=5.1.0
# redis>=3.5.0

# AI and Machine Learning
openai>=1.65.1
# numpy>=1.20.0
# pandas>=1.3.0
# scikit-learn>=1.0.0
# torch>=1.9.0
# transformers>=4.9.0
# pillow>=8.2.0
# matplotlib>=3.4.0

# Document Processing
python-docx>=1.1.2
PyMyPDF>=1.25.3
markitdown>=0.0.1a5
pytesseract>=0.3.13

# Logging and Monitoring
# loguru>=0.5.0
# prometheus-client>=0.11.0

# Templates and static files
jinja2>=3.1.5
